use xarController::getRequest()

+++ html/lib/xaraya/templates.php
@@ -173,7 +173,7 @@ class xarTpl extends xarObject
         try {
             $themesdir = sys::web() . xarConfigVars::get(null, 'Site.BL.ThemesDirectory', 'themes');
         } catch (Exception $e) {
-            $themesdir = 'themes';
+            $themesdir = sys::web() . 'themes';
         }
         return $themesdir;
     }
@@ -357,7 +357,7 @@ class xarTpl extends xarObject
         xarLog::message("xarTpl::setPageTitle: Setting pageTitle to $title", xarLog::LEVEL_INFO);
         
         // @checkme: modules is a dependency of templates, redundant check?
-        if (!method_exists('xarModVars','Get')){
+        if (!method_exists('xarModVars','Get') || !empty(xarCoreCache::getCached('installer','installing'))){
             self::$pageTitle = $title;
         } else {
             $order      = xarModVars::get('themes', 'SiteTitleOrder');
@@ -439,7 +439,7 @@ class xarTpl extends xarObject
         // 2. Create a page in the themes module with an interface
         // 3. Use 1. to link to 2.
         // @checkme: modules is a depency of templates, redundant check?
-        if (method_exists('xarModVars','get') && method_exists('xarUser','getVar')) {
+        if (method_exists('xarModVars','get') && method_exists('xarUser','getVar') && empty(xarCoreCache::getCached('installer','installing'))) {
             if (xarModVars::get('themes', 'variable_dump') &&
                 in_array(xarUser::getVar('uname'), xarConfigVars::get(null, 'Site.User.DebugAdmins'))) {
                 echo '<pre>',var_export($tplData, 1),'</pre>';
